updated the UTs
